DevAI_bk/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
pymongo==4.6.0

# Git Integration - THIS WAS MISSING!
GitPython==3.1.40

# Data Processing
pandas==2.2.0
numpy>=1.24.0

# Visualization
plotly==5.18.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# RAG Pipeline & AI
chromadb>=0.4.18
sentence-transformers>=2.2.2
google-generativeai>=0.3.0
openai>=1.6.1

# Additional Utilities
requests>=2.31.0
typing-extensions>=4.8.0